Moves devtools search UI to the main header
